third_party/WebKit/LayoutTests/fast/viewport/viewport-75-expected.txt

CONSOLE MESSAGE: line 3: The key "x;" is not recognized and ignored.
CONSOLE MESSAGE: line 3: Note that ';' is not a key-value pair separator. The list should be comma-separated.
ALERT: viewport size 64x70.4 scale 5 with limits [5, 5] and userScalable true
